Added code that removes weird paths from filenames in linker errors.
All cached object files' names included info about the architecture that
they were built against. This applies to libraries, personal libraries
and core libraries (such as the arduino core library for Arduino Uno).
In case of a linking error, these are removed, since they were rather
confusing than helpful. Additionally, depending on whether the object
file belongs to the project, or a library a text is added to the
output making this clear.